[feat] salePostImage: 판매글 이미지 개별 관리를 위해 response의 도메인 이동 및 request DTO 구현
- 도메인 이동: .domain.salepost -> .domain.salepostimage
